создал параметры командной строки для watcher

+++ Watcher/Watcher.cs
@@ -9,12 +9,14 @@ namespace Watcher
         private readonly IMessageSender<BusMessage> _sender;
         private readonly IDirectoryWatcher _directoryWatcher;
         private readonly IFileManager _fileManager;
+        private readonly bool _shouldDelete;
 
-        public Watcher(IMessageSender<BusMessage> sender, IDirectoryWatcher directoryWatcher, IFileManager fileManager)
+        public Watcher(IMessageSender<BusMessage> sender, IDirectoryWatcher directoryWatcher, IFileManager fileManager, bool? isDelete)
         {
             _sender = sender ?? throw new ArgumentNullException(nameof(sender));
             _directoryWatcher = directoryWatcher ?? throw new ArgumentNullException(nameof(directoryWatcher));
             _fileManager = fileManager ?? throw new ArgumentNullException(nameof(fileManager));
+            if (isDelete != null) _shouldDelete = (bool) isDelete;
             _directoryWatcher.NewFile += OnNewFile;
             _directoryWatcher.BeginWatch();
         }
@@ -33,7 +35,7 @@ namespace Watcher
                 Content = content, 
                 FileName = fileName
             });
-            _fileManager.Delete(fullPath);
+            if (_shouldDelete) _fileManager.Delete(fullPath);
         }
 
         private static bool IsAllowed(string fullPath)
